fixed missing changes

+++ server/src/main/kotlin/com/kamelia/hedera/rest/user/DTO.kt
@@ -17,7 +17,7 @@ fun User.toRepresentationDTO() = UserRepresentationDTO(
     role,
     enabled,
     currentDiskQuota,
-    if(maximumDiskQuota > 0.toLong()) currentDiskQuota.toDouble() / maximumDiskQuota.toDouble() else 0.toDouble(),
+    if (maximumDiskQuota > 0L) currentDiskQuota.toDouble() / maximumDiskQuota.toDouble() else 0.0,
     maximumDiskQuota,
     createdAt.toString(),
 )
